update switch issues
