+++ apps/web/src/wallet/WalletProvider.tsx
@@ -1,18 +1,15 @@
 import { isInBinance } from '@binance/w3w-utils'
 import dynamic from 'next/dynamic'
 import { useRouter } from 'next/router'
-import { createW3WWagmiConfig, createWagmiConfig } from 'utils/wagmi'
-import { useAtom, useSetAtom } from 'jotai'
+import { useAtom } from 'jotai'
 import { usePrivy } from '@privy-io/react-auth'
 import { atomWithStorage } from 'jotai/utils'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { WagmiProvider as PrivyWagmiProvider } from '@privy-io/wagmi'
-import { useAccountEffect } from 'wagmi'
 import { SOLANA_SUPPORTED_PATH } from './solana.config'
 import { W3WConfigProvider } from './W3WConfigContext'
-import { useRequestChainUpdate } from './hook/useRequestChainUpdate'
 import { useSyncWagmiState } from './hook/useSyncWagmiState'
-import { switchChainRequestAtom } from './atoms/switchChainRequestAtom'
+import { useWagmiConfig } from './hook/useWagmiConfig'
 
 interface WalletProviderProps {
   reconnectOnMount?: boolean
@@ -94,40 +91,12 @@ const usePrivyProvider = () => {
 
 export const WalletProvider = (props: WalletProviderProps) => {
   const { children } = props
-  const [ready, setReady] = useState(false)
   const router = useRouter()
-  const updateRequestChain = useSetAtom(switchChainRequestAtom)
   usePrivyProvider()
-  const wagmiConfig = useMemo(
-    () => (typeof window !== 'undefined' && isInBinance() ? createW3WWagmiConfig() : createWagmiConfig()),
-    [ready],
-  )
 
-  useEffect(() => {
-    window.addEventListener('eip6963:announceProvider', (event: any) => {
-      const { provider } = event.detail
-      eip6963Providers.push(provider)
-    })
-    window.dispatchEvent(new Event('eip6963:requestProvider'))
-    setTimeout(() => {
-      setReady(true)
-    })
-  }, [])
+  const wagmiConfig = useWagmiConfig()
 
-  useAccountEffect({
-    config: wagmiConfig,
-    onConnect: (data) => {
-      const { chainId, isReconnected } = data
-      if (!isReconnected) {
-        updateRequestChain((prev) => ({
-          ...prev,
-          chainId,
-          from: 'wagmi',
-        }))
-      }
-    },
-  })
-  if (!ready) {
+  if (!wagmiConfig) {
     return null // or a loading spinner
   }
 
@@ -144,7 +113,6 @@ export const WalletProvider = (props: WalletProviderProps) => {
 }
 
 const Sync = () => {
-  useRequestChainUpdate()
   useSyncWagmiState()
   return null
 }
